+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added more information to model cards for multiple choice models (`mc-roberta-commonsenseqa`,
 `mc-roberta-piqa`, and `mc-roberta-swag`).
 
+### Fixed
+
+- Fixed many training configs to work out-of-the box. These include the configs for `bart_cnn_dm`, `swag`, `bidaf`, `bidaf_elmo`,
+  `naqanet`, and `qanet`.
+
 ## [v1.2.0rc1](https://github.com/allenai/allennlp-models/releases/tag/v1.2.0rc1) - 2020-10-22
 
 ### Added